master-degree/processo-sviluppo-software/assignment-1/urluckynum/requirements.txt
prospector
bandit
pytest
mysql-connector-python
setuptools
wheel
twine
pdoc